Итог. задача 2

+++ src/FinalTask2.java
@@ -5,18 +5,18 @@ public class FinalTask2 {
     public static void main(String[] arg) {
         int arr = 0;
         int counter, counter_j, var = 0;
-        double temp;
+        int temp;
 
         Scanner in = new Scanner(System.in);
         System.out.print("Введите желаемый размер массива: ");
         arr = in.nextInt();
 
-        double array[] = new double[arr];
+        int array[] = new int[arr];
         for (counter = 0; counter < arr; counter++) {
             Scanner on = new Scanner(System.in);
             var = counter + 1;
             System.out.print("Введите ( " + var + ") элемент массива: ");
-            array[counter] = in.nextDouble();
+            array[counter] = in.nextInt();
         }
         //Sort
         int x = 0;
@@ -44,6 +44,7 @@ public class FinalTask2 {
             }
             x++;
         }
+
         while (x < y);
         //Вывод консоль
         System.out.println("Отсортированный массив: " + Arrays.toString(array));
